+++ src/Controller/AppController.php
@@ -23,7 +23,6 @@
 namespace MeCms\Controller;
 
 use App\Controller\AppController as BaseController;
-use Cake\Core\Configure;
 
 /**
  * Application controller class
@@ -165,6 +164,7 @@ class AppController extends BaseController {
 	 */
 	public function initialize() {
 		//Loads components
+		$this->loadComponent('Cookie');
 		$this->loadComponent('MeCms.Auth');
         $this->loadComponent('MeTools.Flash');
         $this->loadComponent('RequestHandler');