+++ src/View/View/AppView.php
@@ -24,6 +24,7 @@ namespace MeCms\View\View;
 
 use App\View\AppView as BaseView;
 use Cake\Routing\Router;
+use Cake\Cache\Cache;
 
 /**
  * Application view class
@@ -55,6 +56,7 @@ class AppView extends BaseView {
 	 * Gets the title for layou
 	 * @return string Title
 	 * @uses title
+	 * @uses viewVars
 	 */
 	protected function _getTitleForLayout() {
 		if(!empty($this->title))
@@ -94,8 +96,9 @@ class AppView extends BaseView {
 		$this->loadHelper('MeTools.Library');
 		$this->loadHelper('MeTools.Thumb');
 		$this->loadHelper('MeTools.Paginator');
-		$this->loadHelper('MeCms.Auth');
 		$this->loadHelper('MeTools.Recaptcha');
+		$this->loadHelper('MeCms.Auth');
+		$this->loadHelper('MeCms.Widget');
     }
 	
 	/**
@@ -104,7 +107,7 @@ class AppView extends BaseView {
 	 * @param string|NULL $layout Layout to use
 	 * @return string|NULL Rendered content or NULL if content already rendered and returned earlier
 	 * @see http://api.cakephp.org/3.1/class-Cake.View.View.html#_render
-     * @throws Cake\Core\Exception\Exception
+	 * @uses App\View\AppView::render()
 	 * @uses layout
 	 * @uses theme
 	 */
@@ -126,8 +129,10 @@ class AppView extends BaseView {
 	 * @param string|null $layout Layout name
 	 * @return mixed Rendered output, or false on error
 	 * @see http://api.cakephp.org/3.1/source-class-Cake.View.View.html#477-513
-     * @throws Cake\Core\Exception\Exception
+	 * @uses App\View\AppView::renderLayout()
 	 * @uses MeTools\View\Helper\HtmlHelper::meta()
+	 * @uses MeTools\View\Helper\LibraryHelper::analytics()
+	 * @uses MeTools\View\Helper\LibraryHelper::shareaholic()
 	 * @uses _getTitleForLayout()
 	 * @uses _addFacebookTags()
 	 */
@@ -151,35 +156,5 @@ class AppView extends BaseView {
 		$this->_addFacebookTags();
 		
 		return parent::renderLayout($content, $layout);
-	}
-	
-	/**
-	 * Returns all widgets, reading from configuration
-	 * @return string Html code
-	 * @uses MeTools\Network\Request::isCurrent()
-	 * @uses widget()
-	 */
-	public function allWidgets() {
-		$widgets = config('frontend.widgets.general');
-		
-		if($this->request->isCurrent(['_name' => 'homepage']) && config('frontend.widgets.homepage'))
-			$widgets = config('frontend.widgets.homepage');
-		
-		foreach($widgets as $name => $args)
-			$widgets[$name] = is_array($args) ? $this->widget($name, $args) : $this->widget($args);
-		
-		return implode(PHP_EOL, $widgets);
-	}
-	
-	/**
-	 * Returns a widget
-	 * @param string $name Widget name
-	 * @param array $arguments Widget arguments
-	 * @param array $options Widget options
-	 * @return Cake\View\Cell The cell instance
-	 * @uses Cake\View\Cell::cell()
-	 */
-	public function widget($name, array $arguments = [], array $options = []) {
-		return $this->cell($name, $arguments, $options);
 	}
 }