+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # 2.x branch
 ## 2.1 branch
+### 2.1.9
+* banners have "created" and "modified" fields. Photos have "modified" field;
+* fixed bug for cookies policy functionality;
+* fixed bug, now the preview image of the post is displayed correctly in RSS;
+* added some utility links on the footer;
+* added routes for "posts of today" and "posts of yesterday";
+* improved View classes;
+* the code for backend menus and the frontend widgets has been rewritten;
+* backend topbar and backend menus now use cache.
+
 ### 2.1.8
 * added the cookies policy functionality;
 * tags use space instead of the hyphen;